Make PSU info an explicit structure
The previous method introduced extra complexity, requiring trait
objects, explicit unwind safety, etc, and `unsafe` hacks to avoid people
making their own info structures. By transforming it from a type to a
value, we sidestep that all.

The one thing we lose is the ability to define at compile-time that
we're using a particular supply. At this point, we weren't using that
anyway.